+++ packages-resource/midway-test-inner-component/src/componentA/configuration.ts
@@ -1,0 +1,13 @@
+// src/configuration.ts
+import { Configuration } from '@midwayjs/core';
+import * as DefaultConfig from './config/config.default';
+
+@Configuration({
+  namespace: 'A',
+  importConfigs: [
+    {
+      default: DefaultConfig,
+    },
+  ],
+})
+export class AConfiguration {}