+++ packages-resource/midway-test-inner-component/src/componentB/index.ts
@@ -1,0 +1,2 @@
+export { BConfiguration as Configuration } from './configuration';
+export * from './controller/b.controller';