+++ scripts/rollback/rollback-3.6.0.sh
@@ -1,0 +1,291 @@
+#!/bin/bash
+# timestamp: 1665405442697
+
+npm dist-tag add @midwayjs/faas-typings@3.5.0 latest
+npm dist-tag add @midwayjs/faas-typings@3.6.0 beta
+npm dist-tag add @midwayjs/fc-starter@3.5.3 latest
+npm dist-tag add @midwayjs/fc-starter@3.6.0 beta
+npm dist-tag add @midwayjs/serverless-http-parser@3.5.3 latest
+npm dist-tag add @midwayjs/serverless-http-parser@3.6.0 beta
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.5.3 latest
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.6.0 beta
+npm dist-tag add @midwayjs/axios@3.5.3 latest
+npm dist-tag add @midwayjs/axios@3.6.0 beta
+npm dist-tag add @midwayjs/bootstrap@3.5.3 latest
+npm dist-tag add @midwayjs/bootstrap@3.6.0 beta
+npm dist-tag add @midwayjs/bull-board@3.0.1 latest
+npm dist-tag add @midwayjs/bull-board@3.6.0 beta
+npm dist-tag add @midwayjs/bull@3.0.5 latest
+npm dist-tag add @midwayjs/bull@3.6.0 beta
+npm dist-tag add @midwayjs/cache@3.5.3 latest
+npm dist-tag add @midwayjs/cache@3.6.0 beta
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.5.3 latest
+npm dist-tag add @midwayjs/casbin-redis-adapter@3.6.0 beta
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.5.3 latest
+npm dist-tag add @midwayjs/casbin-typeorm-adapter@3.6.0 beta
+npm dist-tag add @midwayjs/casbin@3.5.4 latest
+npm dist-tag add @midwayjs/casbin@3.6.0 beta
+npm dist-tag add @midwayjs/code-dye@3.5.3 latest
+npm dist-tag add @midwayjs/code-dye@3.6.0 beta
+npm dist-tag add @midwayjs/consul@3.5.3 latest
+npm dist-tag add @midwayjs/consul@3.6.0 beta
+npm dist-tag add @midwayjs/core@3.5.3 latest
+npm dist-tag add @midwayjs/core@3.6.0 beta
+npm dist-tag add @midwayjs/cos@3.5.3 latest
+npm dist-tag add @midwayjs/cos@3.6.0 beta
+npm dist-tag add @midwayjs/cross-domain@3.5.3 latest
+npm dist-tag add @midwayjs/cross-domain@3.6.0 beta
+npm dist-tag add @midwayjs/decorator@3.4.11 latest
+npm dist-tag add @midwayjs/decorator@3.6.0 beta
+npm dist-tag add @midwayjs/express-session@3.5.3 latest
+npm dist-tag add @midwayjs/express-session@3.6.0 beta
+npm dist-tag add @midwayjs/faas@3.5.3 latest
+npm dist-tag add @midwayjs/faas@3.6.0 beta
+npm dist-tag add @midwayjs/grpc@3.5.3 latest
+npm dist-tag add @midwayjs/grpc@3.6.0 beta
+npm dist-tag add @midwayjs/http-proxy@3.5.3 latest
+npm dist-tag add @midwayjs/http-proxy@3.6.0 beta
+npm dist-tag add @midwayjs/i18n@3.5.3 latest
+npm dist-tag add @midwayjs/i18n@3.6.0 beta
+npm dist-tag add @midwayjs/info@3.5.3 latest
+npm dist-tag add @midwayjs/info@3.6.0 beta
+npm dist-tag add @midwayjs/jwt@3.5.3 latest
+npm dist-tag add @midwayjs/jwt@3.6.0 beta
+npm dist-tag add @midwayjs/kafka@3.5.3 latest
+npm dist-tag add @midwayjs/kafka@3.6.0 beta
+npm dist-tag add @midwayjs/mikro@3.5.3 latest
+npm dist-tag add @midwayjs/mikro@3.6.0 beta
+npm dist-tag add @midwayjs/mock@3.5.3 latest
+npm dist-tag add @midwayjs/mock@3.6.0 beta
+npm dist-tag add @midwayjs/mongoose@3.5.3 latest
+npm dist-tag add @midwayjs/mongoose@3.6.0 beta
+npm dist-tag add @midwayjs/oss@3.5.3 latest
+npm dist-tag add @midwayjs/oss@3.6.0 beta
+npm dist-tag add @midwayjs/otel@3.5.3 latest
+npm dist-tag add @midwayjs/otel@3.6.0 beta
+npm dist-tag add @midwayjs/passport@3.5.3 latest
+npm dist-tag add @midwayjs/passport@3.6.0 beta
+npm dist-tag add @midwayjs/process-agent@3.5.3 latest
+npm dist-tag add @midwayjs/process-agent@3.6.0 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@3.5.3 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.6.0 beta
+npm dist-tag add @midwayjs/prometheus@3.5.3 latest
+npm dist-tag add @midwayjs/prometheus@3.6.0 beta
+npm dist-tag add @midwayjs/rabbitmq@3.5.3 latest
+npm dist-tag add @midwayjs/rabbitmq@3.6.0 beta
+npm dist-tag add @midwayjs/redis@3.5.3 latest
+npm dist-tag add @midwayjs/redis@3.6.0 beta
+npm dist-tag add @midwayjs/security@3.5.3 latest
+npm dist-tag add @midwayjs/security@3.6.0 beta
+npm dist-tag add @midwayjs/sequelize@3.5.3 latest
+npm dist-tag add @midwayjs/sequelize@3.6.0 beta
+npm dist-tag add @midwayjs/session@3.5.3 latest
+npm dist-tag add @midwayjs/session@3.6.0 beta
+npm dist-tag add @midwayjs/socketio@3.5.3 latest
+npm dist-tag add @midwayjs/socketio@3.6.0 beta
+npm dist-tag add @midwayjs/static-file@3.5.3 latest
+npm dist-tag add @midwayjs/static-file@3.6.0 beta
+npm dist-tag add @midwayjs/swagger@3.5.3 latest
+npm dist-tag add @midwayjs/swagger@3.6.0 beta
+npm dist-tag add @midwayjs/tablestore@3.5.3 latest
+npm dist-tag add @midwayjs/tablestore@3.6.0 beta
+npm dist-tag add @midwayjs/task@3.5.3 latest
+npm dist-tag add @midwayjs/task@3.6.0 beta
+npm dist-tag add @midwayjs/typegoose@3.5.3 latest
+npm dist-tag add @midwayjs/typegoose@3.6.0 beta
+npm dist-tag add @midwayjs/typeorm@3.5.3 latest
+npm dist-tag add @midwayjs/typeorm@3.6.0 beta
+npm dist-tag add @midwayjs/upload@3.5.3 latest
+npm dist-tag add @midwayjs/upload@3.6.0 beta
+npm dist-tag add @midwayjs/validate@3.5.3 latest
+npm dist-tag add @midwayjs/validate@3.6.0 beta
+npm dist-tag add @midwayjs/version@3.5.3 latest
+npm dist-tag add @midwayjs/version@3.6.0 beta
+npm dist-tag add @midwayjs/view-ejs@3.5.3 latest
+npm dist-tag add @midwayjs/view-ejs@3.6.0 beta
+npm dist-tag add @midwayjs/view-nunjucks@3.5.3 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.6.0 beta
+npm dist-tag add @midwayjs/view@3.5.3 latest
+npm dist-tag add @midwayjs/view@3.6.0 beta
+npm dist-tag add @midwayjs/express@3.5.3 latest
+npm dist-tag add @midwayjs/express@3.6.0 beta
+npm dist-tag add @midwayjs/koa@3.5.3 latest
+npm dist-tag add @midwayjs/koa@3.6.0 beta
+npm dist-tag add @midwayjs/web@3.5.3 latest
+npm dist-tag add @midwayjs/web@3.6.0 beta
+npm dist-tag add @midwayjs/ws@3.5.3 latest
+npm dist-tag add @midwayjs/ws@3.6.0 beta
+tnpm dist-tag add @midwayjs/faas-typings@3.5.0 latest
+tnpm dist-tag add @midwayjs/faas-typings@3.6.0 beta
+tnpm dist-tag add @midwayjs/fc-starter@3.5.3 latest
+tnpm dist-tag add @midwayjs/fc-starter@3.6.0 beta
+tnpm dist-tag add @midwayjs/serverless-http-parser@3.5.3 latest
+tnpm dist-tag add @midwayjs/serverless-http-parser@3.6.0 beta
+tnpm dist-tag add @midwayjs/async-hooks-context-manager@3.5.3 latest
+tnpm dist-tag add @midwayjs/async-hooks-context-manager@3.6.0 beta
+tnpm dist-tag add @midwayjs/axios@3.5.3 latest
+tnpm dist-tag add @midwayjs/axios@3.6.0 beta
+tnpm dist-tag add @midwayjs/bootstrap@3.5.3 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.6.0 beta
+tnpm dist-tag add @midwayjs/bull-board@3.0.1 latest
+tnpm dist-tag add @midwayjs/bull-board@3.6.0 beta
+tnpm dist-tag add @midwayjs/bull@3.0.5 latest
+tnpm dist-tag add @midwayjs/bull@3.6.0 beta
+tnpm dist-tag add @midwayjs/cache@3.5.3 latest
+tnpm dist-tag add @midwayjs/cache@3.6.0 beta
+tnpm dist-tag add @midwayjs/casbin-redis-adapter@3.5.3 latest
+tnpm dist-tag add @midwayjs/casbin-redis-adapter@3.6.0 beta
+tnpm dist-tag add @midwayjs/casbin-typeorm-adapter@3.5.3 latest
+tnpm dist-tag add @midwayjs/casbin-typeorm-adapter@3.6.0 beta
+tnpm dist-tag add @midwayjs/casbin@3.5.4 latest
+tnpm dist-tag add @midwayjs/casbin@3.6.0 beta
+tnpm dist-tag add @midwayjs/code-dye@3.5.3 latest
+tnpm dist-tag add @midwayjs/code-dye@3.6.0 beta
+tnpm dist-tag add @midwayjs/consul@3.5.3 latest
+tnpm dist-tag add @midwayjs/consul@3.6.0 beta
+tnpm dist-tag add @midwayjs/core@3.5.3 latest
+tnpm dist-tag add @midwayjs/core@3.6.0 beta
+tnpm dist-tag add @midwayjs/cos@3.5.3 latest
+tnpm dist-tag add @midwayjs/cos@3.6.0 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.5.3 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.6.0 beta
+tnpm dist-tag add @midwayjs/decorator@3.4.11 latest
+tnpm dist-tag add @midwayjs/decorator@3.6.0 beta
+tnpm dist-tag add @midwayjs/express-session@3.5.3 latest
+tnpm dist-tag add @midwayjs/express-session@3.6.0 beta
+tnpm dist-tag add @midwayjs/faas@3.5.3 latest
+tnpm dist-tag add @midwayjs/faas@3.6.0 beta
+tnpm dist-tag add @midwayjs/grpc@3.5.3 latest
+tnpm dist-tag add @midwayjs/grpc@3.6.0 beta
+tnpm dist-tag add @midwayjs/http-proxy@3.5.3 latest
+tnpm dist-tag add @midwayjs/http-proxy@3.6.0 beta
+tnpm dist-tag add @midwayjs/i18n@3.5.3 latest
+tnpm dist-tag add @midwayjs/i18n@3.6.0 beta
+tnpm dist-tag add @midwayjs/info@3.5.3 latest
+tnpm dist-tag add @midwayjs/info@3.6.0 beta
+tnpm dist-tag add @midwayjs/jwt@3.5.3 latest
+tnpm dist-tag add @midwayjs/jwt@3.6.0 beta
+tnpm dist-tag add @midwayjs/kafka@3.5.3 latest
+tnpm dist-tag add @midwayjs/kafka@3.6.0 beta
+tnpm dist-tag add @midwayjs/mikro@3.5.3 latest
+tnpm dist-tag add @midwayjs/mikro@3.6.0 beta
+tnpm dist-tag add @midwayjs/mock@3.5.3 latest
+tnpm dist-tag add @midwayjs/mock@3.6.0 beta
+tnpm dist-tag add @midwayjs/mongoose@3.5.3 latest
+tnpm dist-tag add @midwayjs/mongoose@3.6.0 beta
+tnpm dist-tag add @midwayjs/oss@3.5.3 latest
+tnpm dist-tag add @midwayjs/oss@3.6.0 beta
+tnpm dist-tag add @midwayjs/otel@3.5.3 latest
+tnpm dist-tag add @midwayjs/otel@3.6.0 beta
+tnpm dist-tag add @midwayjs/passport@3.5.3 latest
+tnpm dist-tag add @midwayjs/passport@3.6.0 beta
+tnpm dist-tag add @midwayjs/process-agent@3.5.3 latest
+tnpm dist-tag add @midwayjs/process-agent@3.6.0 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.5.3 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.6.0 beta
+tnpm dist-tag add @midwayjs/prometheus@3.5.3 latest
+tnpm dist-tag add @midwayjs/prometheus@3.6.0 beta
+tnpm dist-tag add @midwayjs/rabbitmq@3.5.3 latest
+tnpm dist-tag add @midwayjs/rabbitmq@3.6.0 beta
+tnpm dist-tag add @midwayjs/redis@3.5.3 latest
+tnpm dist-tag add @midwayjs/redis@3.6.0 beta
+tnpm dist-tag add @midwayjs/security@3.5.3 latest
+tnpm dist-tag add @midwayjs/security@3.6.0 beta
+tnpm dist-tag add @midwayjs/sequelize@3.5.3 latest
+tnpm dist-tag add @midwayjs/sequelize@3.6.0 beta
+tnpm dist-tag add @midwayjs/session@3.5.3 latest
+tnpm dist-tag add @midwayjs/session@3.6.0 beta
+tnpm dist-tag add @midwayjs/socketio@3.5.3 latest
+tnpm dist-tag add @midwayjs/socketio@3.6.0 beta
+tnpm dist-tag add @midwayjs/static-file@3.5.3 latest
+tnpm dist-tag add @midwayjs/static-file@3.6.0 beta
+tnpm dist-tag add @midwayjs/swagger@3.5.3 latest
+tnpm dist-tag add @midwayjs/swagger@3.6.0 beta
+tnpm dist-tag add @midwayjs/tablestore@3.5.3 latest
+tnpm dist-tag add @midwayjs/tablestore@3.6.0 beta
+tnpm dist-tag add @midwayjs/task@3.5.3 latest
+tnpm dist-tag add @midwayjs/task@3.6.0 beta
+tnpm dist-tag add @midwayjs/typegoose@3.5.3 latest
+tnpm dist-tag add @midwayjs/typegoose@3.6.0 beta
+tnpm dist-tag add @midwayjs/typeorm@3.5.3 latest
+tnpm dist-tag add @midwayjs/typeorm@3.6.0 beta
+tnpm dist-tag add @midwayjs/upload@3.5.3 latest
+tnpm dist-tag add @midwayjs/upload@3.6.0 beta
+tnpm dist-tag add @midwayjs/validate@3.5.3 latest
+tnpm dist-tag add @midwayjs/validate@3.6.0 beta
+tnpm dist-tag add @midwayjs/version@3.5.3 latest
+tnpm dist-tag add @midwayjs/version@3.6.0 beta
+tnpm dist-tag add @midwayjs/view-ejs@3.5.3 latest
+tnpm dist-tag add @midwayjs/view-ejs@3.6.0 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@3.5.3 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@3.6.0 beta
+tnpm dist-tag add @midwayjs/view@3.5.3 latest
+tnpm dist-tag add @midwayjs/view@3.6.0 beta
+tnpm dist-tag add @midwayjs/express@3.5.3 latest
+tnpm dist-tag add @midwayjs/express@3.6.0 beta
+tnpm dist-tag add @midwayjs/koa@3.5.3 latest
+tnpm dist-tag add @midwayjs/koa@3.6.0 beta
+tnpm dist-tag add @midwayjs/web@3.5.3 latest
+tnpm dist-tag add @midwayjs/web@3.6.0 beta
+tnpm dist-tag add @midwayjs/ws@3.5.3 latest
+tnpm dist-tag add @midwayjs/ws@3.6.0 beta
+
+# Changes:
+
+#  - @midwayjs/faas-typings: 3.5.0 => 3.6.0
+#  - @midwayjs/fc-starter: 3.5.3 => 3.6.0
+#  - @midwayjs/serverless-http-parser: 3.5.3 => 3.6.0
+#  - @midwayjs/async-hooks-context-manager: 3.5.3 => 3.6.0
+#  - @midwayjs/axios: 3.5.3 => 3.6.0
+#  - @midwayjs/bootstrap: 3.5.3 => 3.6.0
+#  - @midwayjs/bull-board: 3.0.1 => 3.6.0
+#  - @midwayjs/bull: 3.0.5 => 3.6.0
+#  - @midwayjs/cache: 3.5.3 => 3.6.0
+#  - @midwayjs/casbin-redis-adapter: 3.5.3 => 3.6.0
+#  - @midwayjs/casbin-typeorm-adapter: 3.5.3 => 3.6.0
+#  - @midwayjs/casbin: 3.5.4 => 3.6.0
+#  - @midwayjs/code-dye: 3.5.3 => 3.6.0
+#  - @midwayjs/consul: 3.5.3 => 3.6.0
+#  - @midwayjs/core: 3.5.3 => 3.6.0
+#  - @midwayjs/cos: 3.5.3 => 3.6.0
+#  - @midwayjs/cross-domain: 3.5.3 => 3.6.0
+#  - @midwayjs/decorator: 3.4.11 => 3.6.0
+#  - @midwayjs/express-session: 3.5.3 => 3.6.0
+#  - @midwayjs/faas: 3.5.3 => 3.6.0
+#  - @midwayjs/grpc: 3.5.3 => 3.6.0
+#  - @midwayjs/http-proxy: 3.5.3 => 3.6.0
+#  - @midwayjs/i18n: 3.5.3 => 3.6.0
+#  - @midwayjs/info: 3.5.3 => 3.6.0
+#  - @midwayjs/jwt: 3.5.3 => 3.6.0
+#  - @midwayjs/kafka: 3.5.3 => 3.6.0
+#  - @midwayjs/mikro: 3.5.3 => 3.6.0
+#  - @midwayjs/mock: 3.5.3 => 3.6.0
+#  - @midwayjs/mongoose: 3.5.3 => 3.6.0
+#  - @midwayjs/oss: 3.5.3 => 3.6.0
+#  - @midwayjs/otel: 3.5.3 => 3.6.0
+#  - @midwayjs/passport: 3.5.3 => 3.6.0
+#  - @midwayjs/process-agent: 3.5.3 => 3.6.0
+#  - @midwayjs/prometheus-socket-io: 3.5.3 => 3.6.0
+#  - @midwayjs/prometheus: 3.5.3 => 3.6.0
+#  - @midwayjs/rabbitmq: 3.5.3 => 3.6.0
+#  - @midwayjs/redis: 3.5.3 => 3.6.0
+#  - @midwayjs/security: 3.5.3 => 3.6.0
+#  - @midwayjs/sequelize: 3.5.3 => 3.6.0
+#  - @midwayjs/session: 3.5.3 => 3.6.0
+#  - @midwayjs/socketio: 3.5.3 => 3.6.0
+#  - @midwayjs/static-file: 3.5.3 => 3.6.0
+#  - @midwayjs/swagger: 3.5.3 => 3.6.0
+#  - @midwayjs/tablestore: 3.5.3 => 3.6.0
+#  - @midwayjs/task: 3.5.3 => 3.6.0
+#  - @midwayjs/typegoose: 3.5.3 => 3.6.0
+#  - @midwayjs/typeorm: 3.5.3 => 3.6.0
+#  - @midwayjs/upload: 3.5.3 => 3.6.0
+#  - @midwayjs/validate: 3.5.3 => 3.6.0
+#  - @midwayjs/version: 3.5.3 => 3.6.0
+#  - @midwayjs/view-ejs: 3.5.3 => 3.6.0
+#  - @midwayjs/view-nunjucks: 3.5.3 => 3.6.0
+#  - @midwayjs/view: 3.5.3 => 3.6.0
+#  - @midwayjs/express: 3.5.3 => 3.6.0
+#  - @midwayjs/koa: 3.5.3 => 3.6.0
+#  - @midwayjs/web: 3.5.3 => 3.6.0
+#  - @midwayjs/ws: 3.5.3 => 3.6.0