+++ packages/upload/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.1](https://github.com/midwayjs/midway/compare/v3.6.0...v3.6.1) (2022-10-13)
+
+
+### Bug Fixes
+
+* [#2309](https://github.com/midwayjs/midway/issues/2309) ([#2371](https://github.com/midwayjs/midway/issues/2371)) ([919242d](https://github.com/midwayjs/midway/commit/919242d1d7228caa0d450960113396627ce359db))
+
+
+
+
+
+# [3.6.0](https://github.com/midwayjs/midway/compare/v3.5.3...v3.6.0) (2022-10-10)
+
+
+### Features
+
+* add guard ([#2345](https://github.com/midwayjs/midway/issues/2345)) ([1b952a1](https://github.com/midwayjs/midway/commit/1b952a1b09adbb88ff3cff9a2974eb1e37ce53a5))
+
+
+
+
+
 ## [3.5.3](https://github.com/midwayjs/midway/compare/v3.5.2...v3.5.3) (2022-09-25)
 
 **Note:** Version bump only for package @midwayjs/upload